Move the mock props into its own file. Moved the frontendglobals initialization somewhere less obscure
