+++ src/tests/clipperUI/components/annotationInput_tests.tsx
@@ -7,6 +7,7 @@ import {ComponentBase} from "../../../scripts/clipperUI/componentBase";
 
 import {HelperFunctions} from "../../helperFunctions";
 import {MithrilUtils} from "../../mithrilUtils";
+import {MockProps} from "../../mockProps";
 import {TestModule} from "../../testModule";
 
 export class AnnotationInputTests extends TestModule {
@@ -18,7 +19,7 @@ export class AnnotationInputTests extends TestModule {
 
 	protected beforeEach() {
 		this.defaultComponent =
-			<AnnotationInput clipperState={ HelperFunctions.getMockClipperState() } />;
+			<AnnotationInput clipperState={ MockProps.getMockClipperState() } />;
 	}
 
 	protected tests() {
