+++ src/tests/tests.ts
@@ -2,6 +2,19 @@ import {Polyfills} from "../scripts/polyfills";
 
 Polyfills.init();
 
+// TODO move this into an init()
+import {Clipper} from "../scripts/clipperUI/frontEndGlobals";
+import {Communicator} from "../scripts/communicator/communicator";
+import {MockMessageHandler} from "./communicator/mockMessageHandler";
+import {ConsoleLoggerDecorator} from "../scripts/logging/consoleLoggerDecorator";
+import {MockConsole} from "./logging/mockConsole";
+import {ProductionRequirements} from "../scripts/logging/context";
+Clipper.setInjectCommunicator(new Communicator(new MockMessageHandler(), "INJECT_MOCK_COMM"));
+Clipper.setExtensionCommunicator(new Communicator(new MockMessageHandler(), "EXTENSION_MOCK_COMM"));
+Clipper.logger = new ConsoleLoggerDecorator(new MockConsole(), {
+	contextStrategy: new ProductionRequirements()
+});
+
 import "./clipperUI/components/previewViewer/augmentationPreview_tests";
 import "./clipperUI/components/previewViewer/fullPagePreview_tests";
 import "./clipperUI/components/previewViewer/pdfPageViewport_tests";
